ui/cocoa: Implement dpy_mouse_set and dpy_cursor_define
This patch added implementation of dpy_mouse_set and dpy_cursor_define
for cocoa UI, and enabled virtio-gpu simple rendering on macOS.

Due to difficulties in changing macOS cursor position near the frequency
of actual hardware updating, the cursor was drawn on the view using
`-drawRect:` method.

The visibility, content, position and size of the cursor sprite was stored
as ivars.